Conflicts:
	src/core/lib/core_data_model/filtered_tree_model.cpp

+++ src/core/lib/core_serialization/file_data_stream.cpp
@@ -21,7 +21,7 @@ FileDataStream::FileDataStream(const char* path, std::ios::openmode mode)
 void FileDataStream::seek(size_t pos)
 {
 	m_fstream.seekg(pos);
-    m_position = m_fstream.tellg();
+	m_position = static_cast<size_t>( m_fstream.tellg() );
 }
 
 size_t FileDataStream::pos() const
@@ -43,7 +43,7 @@ size_t FileDataStream::readRaw(void * o_Data, size_t length)
 {
 	size_t cur = pos();
 	m_fstream.read(static_cast<char*>(o_Data), length);
-    m_position = m_fstream.tellg();
+	m_position = static_cast<size_t>( m_fstream.tellg() );
 	return pos() - cur;
 }
 
@@ -51,20 +51,20 @@ size_t FileDataStream::writeRaw(const void * data, size_t length)
 {
 	size_t cur = pos();
 	m_fstream.write(static_cast<const char*>(data), length);
-    m_position = m_fstream.tellp();
-    return pos() - cur;
+	m_position = static_cast<size_t>( m_fstream.tellp() );
+	return pos() - cur;
 }
 
 bool FileDataStream::writeValue(const Variant & variant)
 {
 	m_fstream << variant;
-    m_position = m_fstream.tellp();
+	m_position = static_cast<size_t>( m_fstream.tellp() );
 	return true;
 }
 
 bool FileDataStream::readValue(Variant & variant)
 {
 	m_fstream >> variant;
-    m_position = m_fstream.tellg();
+	m_position = static_cast<size_t>( m_fstream.tellg() );
 	return true;
 }
